[CMAKE] Override the flags for the Debug CMake configuration
Summary:
This diffs overrides the default flags for the `Debug` build type,
setting the optimization to `-g3 -Og` (or their fallbacks).

Depends on D2653

Test Plan:
  mkdir buildcmake && cd buildcmake
  cmake -GNinja .. -DCMAKE_BUILD_TYPE=Debug
  ninja -v
Check that the `-g3 -Og` flags (or their fallbacks) flags are set in the
compile commands for the CXX compile lines. Check that the `-g -O2`
flags are set in the compile commands for the CC compile lines.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: jasonbcox, teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2644