[CMAKE] Refactor warnings for secp256k1
Summary:
The `-Wshadow` warning is duplicated from the top level CMakeLists.txt
and can therefore be removed.
The `-Wno-nonnull` is not useful as it actually doesn't remove any
warning. This also makes it more consistent with autotools build which
doesn't enforce this warning suppression.

Test Plan:
The warnings should be the same before and after this patch:
  ninja check

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3798