[CMAKE] Use CPack to build packages
Summary:
This diff adds a new `package` build target, which creates one or more
output suitable for installation. For now limited to building a tar.gz
or zip archive of an installation directory, more coming later.

Test Plan:
On all platforms:
  ninja package
Check the archive output is consistent with the items delivered in our
release binaries archives (requires D4229 for the `libbitcoinconsensus`
lib and include).

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D4267