[SECP256K1] Add valgrind check to travis
Summary:
```
As discussed in #687
This adds valgrind check to the repo.

It doesn't run on recovery+ecdh because of the time.
No openssl because of uninitialized mem.
I debated between with and without ASM, but decided with ASM because it
might be more fragile(?).

I wasn't sure if I should pass -DVALGRIND via CFLAGS or CPPFLAGS, it
seems like because this is only C then there shouldn't even be CPPFLAGS
but looks like we use CPPFLAGS in other places for the preprocessor
definitions.

If people are worried about the time it takes we can mark it as
allow_failure although I don't think it's a problem here because there's
only a handful of PRs and they're usually open for weeks.
```

Backport of secp256k1 [[https://github.com/bitcoin-core/secp256k1/pull/690 | PR690]].

Depends on D5429, D5430 and D5431.

Test Plan:
Run the Travis build (see
https://travis-ci.org/Fabcien/secp256k1/builds/658345685).

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5433