[CMAKE] Use a NSIS custom template
Summary:
This allow to use the exact same installation process than the one from
autotools. This uses the same template adapted to work with cmake/cpack.
The installed components are also almost identical.

It is slightly improved over the original template to output a better
selection of the documentation (no `CMakeLists.txt` is included for
example) and a Bitcoin ABC icon for the uninstaller instead of windows
default uninstall icon.

The name and location of the template are requirements from the CPack NSIS generator.
It expects this template to have this exact name and be located in the CMake module path.

Test Plan:
Build for a Windows target:
  ninja package
Run the installer, check the start menu links, run the uninstaller.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4445