[CMAKE] Separate the target for running the unit tests with upgrade
Summary:
Add a new `check-bitcoin-upgrade-activated` target to run the unit tests
with the phonon activation time in the past (there is also a similar new
target for every boost unit tests suite).
This avoids having to reconfigure the build to run the tests.
Note that check-all runs this target as well for now.

Test Plan:
  ninja check-bitcoin # No upgrade
  ninja check-bitcoin-upgrade-activated # With phonon activation
  ninja check-all # Run both

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4799