[SECP256K1] CMake: Fix in-tree build
Summary:
The in-tree build is broken when building with CMake.
This is because the "standard" CMake call pattern has a dual meaning
depending on the context: the path argument is eigher the source root
(which is what we want) **OR** the build root if a build already exists
in this path.
This second case occurs when building in-tree as the source root is also
the build root; as a consequence when configuring the native executables
to generate the build tree to the `native` directory, the root build
files are updated instead.

{meme, src=notabug}

This diff fixes the issue by specifiying both the source root and build
root when calling CMake to configure the native executables. The build
directory is created by CMake if needed, so the initial `mkdir` becomes
useless and is removed.

Test Plan: Build the secp256k1 project both in-tree and out-of-tree.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5097